+++ src/Config/ConfigFormGUI.php
@@ -2,6 +2,7 @@
 
 namespace srag\Plugins\HelpMe\Config;
 
+use ilCheckboxInputGUI;
 use ilEMailInputGUI;
 use ilHelpMePlugin;
 use ilMultiSelectInputGUI;
@@ -36,6 +37,23 @@ class ConfigFormGUI extends ActiveRecordConfigFormGUI {
 	/**
 	 * @inheritdoc
 	 */
+	protected function getValue(/*string*/
+		$key) {
+		switch (true) {
+			case (strpos($key, Config::KEY_RECIPIENT_TEMPLATES . "_") === 0):
+				$template_name = substr($key, strlen(Config::KEY_RECIPIENT_TEMPLATES) + 1);
+
+				return parent::getValue(Config::KEY_RECIPIENT_TEMPLATES)[$template_name];
+
+			default:
+				return parent::getValue($key);
+		}
+	}
+
+
+	/**
+	 * @inheritdoc
+	 */
 	protected function initFields()/*: void*/ {
 		$this->fields = [
 			Config::KEY_RECIPIENT => [
@@ -48,6 +66,13 @@ class ConfigFormGUI extends ActiveRecordConfigFormGUI {
 							Config::KEY_SEND_EMAIL_ADDRESS => [
 								self::PROPERTY_CLASS => ilEMailInputGUI::class,
 								self::PROPERTY_REQUIRED => true
+							],
+							Config::KEY_RECIPIENT_TEMPLATES . "_" . Recipient::SEND_EMAIL => [
+								self::PROPERTY_CLASS => ilSelectInputGUI::class,
+								self::PROPERTY_REQUIRED => true,
+								self::PROPERTY_OPTIONS => [ "" => "" ] + self::notification()->getArrayForSelection(),
+								"setTitle" => $this->txt("template"),
+								"setInfo" => ilNotifications4PluginsPlugin::PLUGIN_NAME
 							]
 						]
 					],
@@ -90,20 +115,33 @@ class ConfigFormGUI extends ActiveRecordConfigFormGUI {
 											Config::KEY_JIRA_ACCESS_TOKEN => [
 												self::PROPERTY_CLASS => ilTextInputGUI::class,
 												self::PROPERTY_REQUIRED => true
-											],
+											]
 										]
 									]
 								]
 							],
+							Config::KEY_RECIPIENT_TEMPLATES . "_" . Recipient::CREATE_JIRA_TICKET => [
+								self::PROPERTY_CLASS => ilSelectInputGUI::class,
+								self::PROPERTY_REQUIRED => true,
+								self::PROPERTY_OPTIONS => [ "" => "" ] + self::notification()->getArrayForSelection(),
+								"setTitle" => $this->txt("template"),
+								"setInfo" => ilNotifications4PluginsPlugin::PLUGIN_NAME
+							]
 						]
 					]
 				]
 			],
-			Config::KEY_TEMPLATE => [
-				self::PROPERTY_CLASS => ilSelectInputGUI::class,
-				self::PROPERTY_REQUIRED => true,
-				self::PROPERTY_OPTIONS => [ "" => "" ] + self::notification()->getArrayForSelection(),
-				"setInfo" => ilNotifications4PluginsPlugin::PLUGIN_NAME
+			Config::KEY_SEND_CONFIRMATION_EMAIL => [
+				self::PROPERTY_CLASS => ilCheckboxInputGUI::class,
+				self::PROPERTY_SUBITEMS => [
+					Config::KEY_RECIPIENT_TEMPLATES . "_" . Config::KEY_SEND_CONFIRMATION_EMAIL => [
+						self::PROPERTY_CLASS => ilSelectInputGUI::class,
+						self::PROPERTY_REQUIRED => true,
+						self::PROPERTY_OPTIONS => [ "" => "" ] + self::notification()->getArrayForSelection(),
+						"setTitle" => $this->txt("template"),
+						"setInfo" => ilNotifications4PluginsPlugin::PLUGIN_NAME
+					]
+				]
 			],
 			Config::KEY_PRIORITIES => [
 				self::PROPERTY_CLASS => ilTextInputGUI::class,
@@ -131,8 +169,19 @@ class ConfigFormGUI extends ActiveRecordConfigFormGUI {
 	 */
 	protected function storeValue(/*string*/
 		$key, $value)/*: void*/ {
-		switch ($key) {
-			case Config::KEY_ROLES:
+		switch (true) {
+			case (strpos($key, Config::KEY_RECIPIENT_TEMPLATES . "_") === 0):
+				$template_name = substr($key, strlen(Config::KEY_RECIPIENT_TEMPLATES) + 1);
+
+				$template_names = $this->getValue(Config::KEY_RECIPIENT_TEMPLATES);
+
+				$template_names[$template_name] = $value;
+
+				$key = Config::KEY_RECIPIENT_TEMPLATES;
+				$value = $template_names;
+				break;
+
+			case ($key === Config::KEY_ROLES):
 				array_shift($value);
 
 				$value = array_map(function (string $role_id): int {