+++ src/Recipient/Recipient.php
@@ -9,9 +9,11 @@ use srag\ActiveRecordConfig\HelpMe\Exception\ActiveRecordConfigException;
 use srag\DIC\HelpMe\DICTrait;
 use srag\DIC\HelpMe\Exception\DICException;
 use srag\HelpMe\Exception\HelpMeException;
+use srag\Plugins\HelpMe\Config\Config;
 use srag\Plugins\HelpMe\Support\Support;
 use srag\Plugins\HelpMe\Support\SupportGUI;
 use srag\Plugins\HelpMe\Utils\HelpMeTrait;
+use srag\Plugins\Notifications4Plugins\Utils\Notifications4PluginsTrait;
 
 /**
  * Class Recipient
@@ -24,6 +26,7 @@ abstract class Recipient {
 
 	use DICTrait;
 	use HelpMeTrait;
+	use Notifications4PluginsTrait;
 	const SEND_EMAIL = "send_email";
 	const CREATE_JIRA_TICKET = "create_jira_ticket";
 	const PLUGIN_CLASS_NAME = ilHelpMePlugin::class;
@@ -82,38 +85,72 @@ abstract class Recipient {
 	 * @throws phpmailerException
 	 */
 	protected function sendConfirmationMail()/*: void*/ {
-		$mailer = new ilMimeMail();
+		if (Config::getField(Config::KEY_SEND_CONFIRMATION_EMAIL)) {
+			$mailer = new ilMimeMail();
 
-		$mailer->From(self::dic()->mailMimeSenderFactory()->system());
+			$mailer->From(self::dic()->mailMimeSenderFactory()->system());
 
-		$mailer->To($this->support->getEmail());
+			$mailer->To($this->support->getEmail());
 
-		$mailer->Subject(self::plugin()->translate("confirmation", SupportGUI::LANG_MODULE_SUPPORT) . ": " . $this->support->getSubject());
+			$mailer->Subject($this->getSubject(Config::KEY_SEND_CONFIRMATION_EMAIL));
 
-		$mailer->Body($this->support->getBody());
+			$mailer->Body($this->getBody(Config::KEY_SEND_CONFIRMATION_EMAIL));
 
-		foreach ($this->support->getScreenshots() as $screenshot) {
-			$mailer->Attach($screenshot->getPath(), $screenshot->getMimeType(), "attachment", $screenshot->getName());
-		}
+			foreach ($this->support->getScreenshots() as $screenshot) {
+				$mailer->Attach($screenshot->getPath(), $screenshot->getMimeType(), "attachment", $screenshot->getName());
+			}
 
-		if (!$mailer->Send()) {
-			throw new HelpMeException("Mailer returns not true");
+			if (!$mailer->Send()) {
+				throw new HelpMeException("Mailer returns not true");
+			}
 		}
 	}
 
 
 	/**
-	 * @return Support
+	 * @param string $template_name
+	 *
+	 * @return string
 	 */
-	public function getSupport(): Support {
-		return $this->support;
+	public function getSubject(string $template_name): string {
+		$notification = self::notification()->getNotificationByName(Config::getField(Config::KEY_RECIPIENT_TEMPLATES)[$template_name]);
+
+		return self::parser()->parseSubject(self::parser()->getParserForNotification($notification), $notification, [
+			"support" => $this->support
+		]);
 	}
 
 
 	/**
-	 * @param Support $support
+	 * @param string $template_name
+	 *
+	 * @return string
 	 */
-	public function setSupport(Support $support)/*: void*/ {
-		$this->support = $support;
+	public function getBody(string $template_name): string {
+		$notification = self::notification()->getNotificationByName(Config::getField(Config::KEY_RECIPIENT_TEMPLATES)[$template_name]);
+
+		$fields_ = [
+			"project" => $this->support->getProject()->getProjectName() . " (" . $this->support->getProject()->getProjectKey() . ")",
+			"title" => $this->support->getTitle(),
+			"name" => $this->support->getName(),
+			"login" => $this->support->getLogin(),
+			"email" => $this->support->getEmail(),
+			"phone" => $this->support->getPhone(),
+			"priority" => $this->support->getPriority(),
+			"description" => $this->support->getDescription(),
+			"reproduce_steps" => $this->support->getReproduceSteps(),
+			"system_infos" => $this->support->getSystemInfos(),
+			"datetime" => $this->support->getFormatedTime()
+		];
+
+		$fields = [];
+		foreach ($fields_ as $key => $value) {
+			$fields[self::plugin()->translate($key, SupportGUI::LANG_MODULE_SUPPORT)] = $value;
+		}
+
+		return self::parser()->parseText(self::parser()->getParserForNotification($notification), $notification, [
+			"support" => $this->support,
+			"fields" => $fields
+		]);
 	}
 }