+++ src/Recipient/RecipientCreateJiraTicket.php
@@ -80,7 +80,8 @@ class RecipientCreateJiraTicket extends Recipient {
 	 */
 	protected function createJiraTicket()/*: void*/ {
 		$issue_key = $this->jira_curl->createJiraIssueTicket($this->support->getProject()->getProjectKey(), $this->support->getProject()
-			->getProjectIssueType(), $this->support->getSubject(), $this->support->getBody(), $this->support->getProject()->getProjectFixVersion());
+			->getProjectIssueType(), $this->getSubject(self::CREATE_JIRA_TICKET), $this->getBody(self::CREATE_JIRA_TICKET), $this->support->getProject()
+			->getProjectFixVersion());
 
 		$this->issue_key = $issue_key;
 	}