+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [x]
+- Config confirmation mail
+- Supports for each recipient a differnt template
+
 ## [3.3.0]
 - Use Notifications4Plugins template
 - Using some new ILIAS 5.3 UI's